+++ s/utils/idb/range.ts
@@ -1,9 +1,10 @@
 
+import {is} from "@e280/stz"
 import {Scan} from "../../types.js"
 
 export function idbRange({start, end}: Scan) {
-	if (start && end) return IDBKeyRange.bound(start, end)
-	if (start) return IDBKeyRange.lowerBound(start)
-	if (end) return IDBKeyRange.upperBound(end)
+	if (is.happy(start) && is.happy(end)) return IDBKeyRange.bound(start, end, false, true)
+	if (is.happy(start)) return IDBKeyRange.lowerBound(start, false)
+	if (is.happy(end)) return IDBKeyRange.upperBound(end, true)
 }