+++ s/utils/idb/wait.ts
@@ -1,5 +1,5 @@
 
-export async function idbWait<T>(request: IDBRequest) {
+export async function idbWait<T>(request: IDBRequest<T>) {
 	return new Promise<T>((resolve, reject) => {
 		request.onsuccess = () => resolve(request.result)
 		request.onerror = () => reject(request.error)