+++ s/utils/lex-sort.ts
@@ -1,0 +1,5 @@
+
+export const lexSort = (a: string, b: string) => (
+	a < b ? -1 : a > b ? 1 : 0
+)
+