+++ s/utils/scan-match.ts
@@ -2,8 +2,8 @@
 import {Scan} from "../types.js"
 
 export function scanMatch(key: string, {start, end}: Scan) {
-	if (start && key < start) return false
-	if (end && key > end) return false
+	if (start !== undefined && key < start) return false
+	if (end !== undefined && key >= end) return false
 	return true
 }